requirements.txt
numpy==1.24.3
matplotlib==3.7.1
torch==1.8.0
torchvision==0.9.0
tqdm==4.65.0
Pillow==10.2.0
opencv-python==4.7.0.72
scipy==1.10.1
scikit-image==0.21.0
scikit-learn==1.3.2
einops== 0.6.1
tensorboardX==2.2
thop== 0.1.1